# Improve consistent initialization speed and robustness