--- a/pybamm/solvers/idaklu_solver.py
+++ b/pybamm/solvers/idaklu_solver.py
@@ -54,7 +54,7 @@ class IDAKLUSolver(pybamm.BaseSolver):
     rtol : float, optional
         The relative tolerance for the solver (default is 1e-6).
     atol : float, optional
-        The absolute tolerance for the solver (default is 1e-6).
+        The absolute tolerance for the solver (default is 1e-4).
     root_method : str or pybamm algebraic solver class, optional
         The method to use to find initial conditions (for DAE solvers).
         If a solver class, must be an algebraic solver class.
@@ -280,26 +280,6 @@ class IDAKLUSolver(pybamm.BaseSolver):
         if isinstance(y0, casadi.DM):
             y0 = y0.full()
         y0 = y0.flatten()
-
-        y0S = model.y0S
-        # only casadi solver needs sensitivity ics
-        if model.convert_to_format != "casadi":
-            y0S = None
-            if self.output_variables and not (
-                model.convert_to_format == "jax"
-                and self._options["jax_evaluator"] == "iree"
-            ):
-                raise pybamm.SolverError(
-                    "output_variables can only be specified "
-                    'with convert_to_format="casadi", or convert_to_format="jax" '
-                    'with jax_evaluator="iree"'
-                )  # pragma: no cover
-        if y0S is not None:
-            if isinstance(y0S, casadi.DM):
-                y0S = (y0S,)
-
-            y0S = (x.full() for x in y0S)
-            y0S = [x.flatten() for x in y0S]
 
         if ics_only:
             return base_set_up_return
@@ -535,13 +515,10 @@ class IDAKLUSolver(pybamm.BaseSolver):
                 for i, dFdp_i in enumerate(dFdp.values()):
                     resvalS[i][:] = dFdy @ yS[i] - dFdyd @ ypS[i] + dFdp_i
 
-        try:
-            atol = model.atol
-        except AttributeError:
-            atol = self.atol
+        atol = getattr(model, "atol", self.atol)
+        atol = self._check_atol_type(atol, y0.size)
 
         rtol = self.rtol
-        atol = self._check_atol_type(atol, y0.size)
 
         if model.convert_to_format == "casadi" or (
             model.convert_to_format == "jax"
@@ -858,72 +835,22 @@ class IDAKLUSolver(pybamm.BaseSolver):
         t_eval : numeric type
             The times at which to compute the solution
         inputs_dict : dict, optional
-            Any input parameters to pass to the model when solving
+            Any input parameters to pass to the model when solving.
         """
         inputs_dict = inputs_dict or {}
         # stack inputs
         if inputs_dict:
-            inputs_dict_keys = list(inputs_dict.keys())  # save order
             arrays_to_stack = [np.array(x).reshape(-1, 1) for x in inputs_dict.values()]
             inputs = np.vstack(arrays_to_stack)
         else:
             inputs = np.array([[]])
-            inputs_dict_keys = []
 
-        # do this here cause y0 is set after set_up (calc consistent conditions)
-        y0 = model.y0
-        if isinstance(y0, casadi.DM):
-            y0 = y0.full()
-        y0 = y0.flatten()
+        y0full = model.y0full
+        ydot0full = model.ydot0full
 
-        y0S = model.y0S
-        if (
-            model.convert_to_format == "jax"
-            and self._options["jax_evaluator"] == "iree"
-        ):
-            if y0S is not None:
-                pybamm.demote_expressions_to_32bit = True
-                # preserve order of inputs
-                y0S = self._demote_64_to_32(
-                    np.concatenate([y0S[k] for k in inputs_dict_keys]).flatten()
-                )
-                y0full = self._demote_64_to_32(np.concatenate([y0, y0S]).flatten())
-                ydot0S = self._demote_64_to_32(np.zeros_like(y0S))
-                ydot0full = self._demote_64_to_32(
-                    np.concatenate([np.zeros_like(y0), ydot0S]).flatten()
-                )
-                pybamm.demote_expressions_to_32bit = False
-            else:
-                y0full = y0
-                ydot0full = np.zeros_like(y0)
-        else:
-            # only casadi solver needs sensitivity ics
-            if model.convert_to_format != "casadi":
-                y0S = None
-            if y0S is not None:
-                if isinstance(y0S, casadi.DM):
-                    y0S = (y0S,)
-
-                y0S = (x.full() for x in y0S)
-                y0S = [x.flatten() for x in y0S]
-
-            # solver works with ydot0 set to zero
-            ydot0 = np.zeros_like(y0)
-            if y0S is not None:
-                ydot0S = [np.zeros_like(y0S_i) for y0S_i in y0S]
-                y0full = np.concatenate([y0, *y0S])
-                ydot0full = np.concatenate([ydot0, *ydot0S])
-            else:
-                y0full = y0
-                ydot0full = ydot0
-
-        try:
-            atol = model.atol
-        except AttributeError:
-            atol = self.atol
-
+        atol = getattr(model, "atol", self.atol)
+        atol = self._check_atol_type(atol, y0full.size)
         rtol = self.rtol
-        atol = self._check_atol_type(atol, y0.size)
 
         timer = pybamm.Timer()
         if model.convert_to_format == "casadi" or (
@@ -939,8 +866,8 @@ class IDAKLUSolver(pybamm.BaseSolver):
         else:
             sol = idaklu.solve_python(
                 t_eval,
-                y0,
-                ydot0,
+                y0full,
+                ydot0full,
                 self._setup["resfn"],
                 self._setup["jac_class"].jac_res,
                 self._setup["sensfn"],
@@ -963,9 +890,8 @@ class IDAKLUSolver(pybamm.BaseSolver):
             "number_of_sensitivity_parameters"
         ]
         sensitivity_names = self._setup["sensitivity_names"]
-        t = sol.t
-        number_of_timesteps = t.size
-        number_of_states = y0.size
+        number_of_timesteps = sol.t.size
+        number_of_states = model.len_rhs_and_alg
         if self.output_variables:
             # Substitute empty vectors for state vector 'y'
             y_out = np.zeros((number_of_timesteps * number_of_states, 0))
@@ -994,63 +920,216 @@ class IDAKLUSolver(pybamm.BaseSolver):
             termination = "event"
         else:
             raise pybamm.SolverError("idaklu solver failed")
+
         newsol = pybamm.Solution(
             sol.t,
             np.transpose(y_out),
             model,
             inputs_dict,
-            np.array([t[-1]]),
+            np.array([sol.t[-1]]),
             np.transpose(y_out[-1])[:, np.newaxis],
             termination,
             sensitivities=yS_out,
         )
         newsol.integration_time = integration_time
-        if self.output_variables:
-            # Populate variables and sensititivies dictionaries directly
-            number_of_samples = sol.y.shape[0] // number_of_timesteps
-            sol.y = sol.y.reshape((number_of_timesteps, number_of_samples))
-            startk = 0
-            for var in self.output_variables:
-                # ExplicitTimeIntegral's are not computed as part of the solver and
-                # do not need to be converted
-                if isinstance(
-                    model.variables_and_events[var], pybamm.ExplicitTimeIntegral
-                ):
-                    continue
-                if model.convert_to_format == "casadi":
-                    len_of_var = (
-                        self._setup["var_fcns"][var](0.0, 0.0, 0.0).sparsity().nnz()
-                    )
-                    base_variables = [self._setup["var_fcns"][var]]
-                elif (
-                    model.convert_to_format == "jax"
-                    and self._options["jax_evaluator"] == "iree"
-                ):
-                    idx = self.output_variables.index(var)
-                    len_of_var = self._setup["var_idaklu_fcns"][idx].nnz
-                    base_variables = [self._setup["var_idaklu_fcns"][idx]]
-                else:  # pragma: no cover
-                    raise pybamm.SolverError(
-                        "Unsupported evaluation engine for convert_to_format="
-                        + f"{model.convert_to_format} "
-                        + f"(jax_evaluator={self._options['jax_evaluator']})"
-                    )
-                newsol._variables[var] = pybamm.ProcessedVariableComputed(
-                    [model.variables_and_events[var]],
-                    base_variables,
-                    [sol.y[:, startk : (startk + len_of_var)]],
-                    newsol,
+        if not self.output_variables:
+            return newsol
+
+        # Populate variables and sensititivies dictionaries directly
+        number_of_samples = sol.y.shape[0] // number_of_timesteps
+        sol.y = sol.y.reshape((number_of_timesteps, number_of_samples))
+        startk = 0
+        for var in self.output_variables:
+            # ExplicitTimeIntegral's are not computed as part of the solver and
+            # do not need to be converted
+            if isinstance(model.variables_and_events[var], pybamm.ExplicitTimeIntegral):
+                continue
+            if model.convert_to_format == "casadi":
+                len_of_var = (
+                    self._setup["var_fcns"][var](0.0, 0.0, 0.0).sparsity().nnz()
                 )
-                # Add sensitivities
-                newsol[var]._sensitivities = {}
-                if model.calculate_sensitivities:
-                    for paramk, param in enumerate(inputs_dict.keys()):
-                        newsol[var].add_sensitivity(
-                            param,
-                            [sol.yS[:, startk : (startk + len_of_var), paramk]],
-                        )
-                startk += len_of_var
+                base_variables = [self._setup["var_fcns"][var]]
+            elif (
+                model.convert_to_format == "jax"
+                and self._options["jax_evaluator"] == "iree"
+            ):
+                idx = self.output_variables.index(var)
+                len_of_var = self._setup["var_idaklu_fcns"][idx].nnz
+                base_variables = [self._setup["var_idaklu_fcns"][idx]]
+            else:  # pragma: no cover
+                raise pybamm.SolverError(
+                    "Unsupported evaluation engine for convert_to_format="
+                    + f"{model.convert_to_format} "
+                    + f"(jax_evaluator={self._options['jax_evaluator']})"
+                )
+            newsol._variables[var] = pybamm.ProcessedVariableComputed(
+                [model.variables_and_events[var]],
+                base_variables,
+                [sol.y[:, startk : (startk + len_of_var)]],
+                newsol,
+            )
+            # Add sensitivities
+            newsol[var]._sensitivities = {}
+            if model.calculate_sensitivities:
+                for paramk, param in enumerate(inputs_dict.keys()):
+                    newsol[var].add_sensitivity(
+                        param,
+                        [sol.yS[:, startk : (startk + len_of_var), paramk]],
+                    )
+            startk += len_of_var
         return newsol
+
+    def _set_consistent_initialization(self, model, time, inputs_dict):
+        """
+        Initialize y0 and ydot0 for the solver. In addition to calculating
+        y0 from BaseSolver, we also calculate ydot0 for semi-explicit DAEs
+
+        Parameters
+        ----------
+        model : :class:`pybamm.BaseModel`
+            The model for which to calculate initial conditions.
+        time : numeric type
+            The time at which to calculate the initial conditions.
+        inputs_dict : dict
+            Any input parameters to pass to the model when solving.
+        """
+
+        # set model.y0
+        super()._set_consistent_initialization(model, time, inputs_dict)
+
+        casadi_format = model.convert_to_format == "casadi"
+        jax_iree_format = (
+            model.convert_to_format == "jax"
+            and self._options["jax_evaluator"] == "iree"
+        )
+
+        y0 = model.y0
+        if isinstance(y0, casadi.DM):
+            y0 = y0.full()
+        y0 = y0.flatten()
+
+        # calculate the time derivatives of the differential equations
+        # for semi-explicit DAEs
+        calc_ydot0 = model.mass_matrix_inv is not None
+        if model.len_rhs > 0 and calc_ydot0:
+            ydot0 = self._rhs_dot_consistent_initialization(model, time, inputs_dict)
+        else:
+            ydot0 = np.zeros_like(y0)
+
+        sensitivity = (model.y0S is not None) and (jax_iree_format or casadi_format)
+        if sensitivity:
+            y0full, ydot0full = self._sensitivity_consistent_initialization(
+                y0, ydot0, model, time, inputs_dict
+            )
+        else:
+            y0full = y0
+            ydot0full = ydot0
+
+        if jax_iree_format:
+            pybamm.demote_expressions_to_32bit = True
+            y0full = self._demote_64_to_32(y0full)
+            ydot0full = self._demote_64_to_32(ydot0full)
+            pybamm.demote_expressions_to_32bit = False
+
+        model.y0full = y0full
+        model.ydot0full = ydot0full
+
+    def _rhs_dot_consistent_initialization(self, model, time, inputs_dict):
+        """
+        Compute the consistent initialization of ydot0 for the differential terms
+        for the solver. If we have a semi-explicit DAE, we can explicitly solve
+        for this value using the consistently initialized y0 vector.
+
+        Parameters
+        ----------
+        model : :class:`pybamm.BaseModel`
+            The model for which to calculate initial conditions.
+        time : numeric type
+            The time at which to calculate the initial conditions.
+        inputs_dict : dict
+            Any input parameters to pass to the model when solving.
+
+        """
+        casadi_format = model.convert_to_format == "casadi"
+
+        inputs_dict = inputs_dict or {}
+        # stack inputs
+        if inputs_dict:
+            arrays_to_stack = [np.array(x).reshape(-1, 1) for x in inputs_dict.values()]
+            inputs = np.vstack(arrays_to_stack)
+        else:
+            inputs = np.array([[]])
+
+        y0 = model.y0
+        if isinstance(y0, casadi.DM):
+            y0 = y0.full()
+        y0 = y0.flatten()
+
+        ydot0 = np.zeros_like(y0)
+        # calculate the time derivatives of the differential equations
+        input_eval = inputs if casadi_format else inputs_dict
+
+        rhs_alg0 = model.rhs_algebraic_eval(time, y0, input_eval)
+        if isinstance(rhs_alg0, casadi.DM):
+            rhs_alg0 = rhs_alg0.full()
+        rhs_alg0 = rhs_alg0.flatten()
+
+        rhs0 = rhs_alg0[: model.len_rhs]
+
+        # for the differential terms, ydot = M^-1 * (rhs)
+        ydot0[: model.len_rhs] = model.mass_matrix_inv.entries @ rhs0
+
+        return ydot0
+
+    def _sensitivity_consistent_initialization(
+        self, y0, ydot0, model, time, inputs_dict
+    ):
+        """
+        Calculate the sensitivity initial conditions of y0 and ydot0 for the solver.
+
+        Parameters
+        ----------
+        y0 : :class:`numpy.array`
+            The initial values of the state vector.
+        ydot0 : :class:`numpy.array`
+            The initial values of the time derivatives of the state vector.
+        model : :class:`pybamm.BaseModel`
+            The model for which to calculate initial conditions.
+        inputs_dict : dict
+            Any input parameters to pass to the model when solving.
+
+        """
+
+        jax_iree_format = (
+            model.convert_to_format == "jax"
+            and self._options["jax_evaluator"] == "iree"
+        )
+
+        inputs_dict = inputs_dict or {}
+        # stack inputs
+        if inputs_dict:
+            inputs_dict_keys = list(inputs_dict.keys())  # save order
+            # arrays_to_stack = [np.array(x).reshape(-1, 1) for x in inputs_dict.values()]
+            # inputs = np.vstack(arrays_to_stack)
+        else:
+            # inputs = np.array([[]])
+            inputs_dict_keys = []
+
+        y0S = model.y0S
+
+        if jax_iree_format:
+            y0S = np.concatenate([y0S[k] for k in inputs_dict_keys])
+        elif isinstance(y0S, casadi.DM):
+            y0S = (y0S,)
+        y0S = (x.full() for x in y0S)
+        y0S = [x.flatten() for x in y0S]
+
+        y0full = np.concatenate([y0, *y0S])
+
+        ydot0S = [np.zeros_like(y0S_i) for y0S_i in y0S]
+        ydot0full = np.concatenate([ydot0, *ydot0S])
+
+        return y0full, ydot0full
 
     def jaxify(
         self,