--- a/tests/integration/test_models/test_full_battery_models/test_lithium_ion/base_lithium_ion_tests.py
+++ b/tests/integration/test_models/test_full_battery_models/test_lithium_ion/base_lithium_ion_tests.py
@@ -133,6 +133,8 @@ class BaseIntegrationTestLithiumIon:
 
     def test_kinetics_asymmetric_butler_volmer(self):
         options = {"intercalation kinetics": "asymmetric Butler-Volmer"}
+        solver = pybamm.CasadiSolver(atol=1e-14, rtol=1e-14)
+
         parameter_values = pybamm.ParameterValues("Marquis2019")
         parameter_values.update(
             {
@@ -141,7 +143,9 @@ class BaseIntegrationTestLithiumIon:
             },
             check_already_exists=False,
         )
-        self.run_basic_processing_test(options, parameter_values=parameter_values)
+        self.run_basic_processing_test(
+            options, parameter_values=parameter_values, solver=solver
+        )
 
     def test_kinetics_linear(self):
         options = {"intercalation kinetics": "linear"}