… CHANGELOG.md
… docs/source/examples/notebooks/solvers/dae-solver.ipynb
… tests/unit/test_solvers/test_base_solver.py
… tests/unit/test_solvers/test_idaklu_solver.py