# Description

The `IDA` solver takes in initial guesses for `y0` and `ydot0`. Currently, we explicitly calculate `y0` using an algebraic solver before passing it to `IDA`. However, we do not estimate `ydot0` - we just set it to `np.zeros_like(y0)` and rely on `IDA` to determine it. This PR makes a few changes to consistent initialization:

1. Analytically compute `ydot0` using the mass matrix inverse. The `IDA` solver is set up only for **implicit** DAEs, so it cannot properly utilize the mass matrix in initializing the system. As a result, it wastes a lot of time re-computing the consistent initialization for `y0` and `ydot0` in `IDACalcIC`. We can improve this by isolating the differential section of our system of equations and explicitly computing `ydot0` for the rhs,
$$\left. \frac{\mathrm{d}y_d^\*}{\mathrm{d}t} \right|_{t_0} = - ( M\_{d,d} ) ^ {-1} \text{rhs}(y_0^\*, t_0) , $$
where the subscript $d$ refers to the equations containing differential terms and $y_0^\*$ is the consistently initialized $y_0$. **This PR reduces the total solve time of a full 1C discharge from 218 ms to 212 ms and reduces the time integrating the system from 45.0 ms to 40.0 ms.** Additionally, this improves the solver robustness because we reduce reliance on `IDACalcIC`, which occasionally fails.
2. Separate `initial_conditions` and `consistent_initialization`. I found that there was not a clear distinction between these ideas in the solver code. In this update, `initial_conditions` refers to the initial values and guesses provided by the solver or from a previous solution. Initial conditions likely will not satisfy the system of equations at `t0`, while `consistent_initialization` refers to the solution that satisfies the system of equations at `t0`. In some cases, we computed the `initial_conditions` more than once for a single solve, which were removed.
3. Various cleanup of the solver codebase. This section will need additional refactoring to unify the API and remove redundancies, but that is beyond the scope of this PR.

## Type of change

Please add a line in the relevant section of [CHANGELOG.md](https://github.com/pybamm-team/PyBaMM/blob/develop/CHANGELOG.md) to document the change (include PR #) - note reverse order of PR #s. If necessary, also add to the list of breaking changes.

- [x] Optimization (back-end change that speeds up the code)

# Key checklist:

- [x] No style issues: `$ pre-commit run` (or `$ nox -s pre-commit`) (see [CONTRIBUTING.md](https://github.com/pybamm-team/PyBaMM/blob/develop/CONTRIBUTING.md#installing-and-using-pre-commit) for how to set this up to run automatically when committing locally, in just two lines of code)
- [x] All tests pass: `$ python run-tests.py --all` (or `$ nox -s tests`)
- [x] The documentation builds: `$ python run-tests.py --doctest` (or `$ nox -s doctests`)

You can run integration tests, unit tests, and doctests together at once, using `$ python run-tests.py --quick` (or `$ nox -s quick`).

## Further checks:

- [x] Code is commented, particularly in hard-to-understand areas
- [x] Tests added that prove fix is effective or that feature works